requirements.txt
h5py==3.10.0
ipykernel==6.29.0
lmfit==1.2.2
matplotlib==3.8.2
neo==0.12.0
numpy==1.26.3
pandas==2.2.0
pyret==0.6.0
PyYAML==6.0.1
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
umap==0.1.1
requests==2.31.0
